Make eip searchable by name
