+++ doc/resource_types.md
@@ -1324,7 +1324,7 @@ Elastic IP resource type.
 ### exist
 
 ```ruby
-describe eip('123.0.456.789') do
+describe eip('my-eip') do
   it { should exist }
 end
 ```
