+++ lib/awspec/stub/eip.rb
@@ -7,7 +7,13 @@ Aws.config[:ec2] = {
         {
           domain: 'vpc',
           public_ip: '123.0.456.789',
-          instance_id: 'i-ec12345a'
+          instance_id: 'i-ec12345a',
+          tags: [
+            {
+              key: 'Name',
+              value: 'my-eip'
+            }
+          ]
         }
       ]
     }
