+++ spec/type/eip_spec.rb
@@ -8,3 +8,9 @@ describe eip('123.0.456.789') do
   it { should be_associated_to('i-ec12345a') }
   it { should belong_to_domain('vpc') }
 end
+
+describe eip('my-eip') do
+  it { should exist }
+  it { should be_associated_to('i-ec12345a') }
+  it { should belong_to_domain('vpc') }
+end
